Refactor: bloom/core
